treat former commands as Modifications with 0-length ChangeInLength
Replaces the Command protocol, which was an alternate ()->Void
expression type that took a different path than Modifications
altogether, with Modifications that don't change the length.

The type-level separation was a tool to avoid mixing Insert/Delete
modifications with each other, and with Select/Modifying.
We can easily express this with a type marker as well:
ChainableModification. That excludes Insert/Delete, so you can mix
Select/Modifying in another block.