make Insert evaluate locations lazily for consistent behavior with Select
This is honestly not needed, because `Insert` can only be used inside
a `Modifying` construct, a block that guards against mutations that
e.g. a NSTextView will not permit. This in turn 'updates' the range
passed into the block itself.

But it also doesn't hurt to make all Modifications behave similarly,
and it may prevent future problems. (For which I have yet to find test cases.)